+++ gDRplots/NEWS.md
@@ -1,3 +1,6 @@
+## gDRplots 0.0.61 - 2025-02-05
+* add values to combo heatmaps and adjust position of plots in the panel
+
 ## gDRplots 0.0.60 - 2025-01-31
 * fix y-axis range for dose-response curves